+++ src/main/java/in/gov/abdm/nmr/db/sql/domain/state/IStateService.java
@@ -1,8 +1,9 @@
 package in.gov.abdm.nmr.db.sql.domain.state;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface IStateService {
 
-    List<StateTO> getStateData(Long countryId);
+    List<StateTO> getStateData(BigInteger countryId);
 }
