+++ src/main/java/in/gov/abdm/nmr/db/sql/domain/state/StateRepository.java
@@ -1,13 +1,15 @@
 package in.gov.abdm.nmr.db.sql.domain.state;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface StateRepository extends JpaRepository<State, Long> {
+public interface StateRepository extends JpaRepository<State, BigInteger> {
 
     @Query(value = "SELECT * FROM state where country=:country", nativeQuery = true)
-    List<State> getState(Long country);
+    List<State> getState(BigInteger country);
+    
 
 }
