+++ src/main/java/in/gov/abdm/nmr/db/sql/domain/sub_district/SubDistrictTO.java
@@ -1,11 +1,13 @@
 package in.gov.abdm.nmr.db.sql.domain.sub_district;
 
+import java.math.BigInteger;
+
 import lombok.Data;
 
 @Data
 public class SubDistrictTO {
 
-    private Long id;
+    private BigInteger id;
     private String name;
     private String code;
 }
